saving work in progress

+++ src/main/java/com/springboot/ecommerce/service/CategoryServiceImpl.java
@@ -1,10 +1,11 @@
 package com.springboot.ecommerce.service;
 
+import com.springboot.ecommerce.assambler.CategoryAssambler;
+import com.springboot.ecommerce.exception.ResourceNotFoundException;
 import com.springboot.ecommerce.model.Category;
 import com.springboot.ecommerce.repository.CategoryRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 
@@ -12,6 +13,7 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService{
 
 
+    private CategoryAssambler categoryAssambler;
     private CategoryRepository categoryRepository;
 
     public CategoryServiceImpl(CategoryRepository categoryRepository) {
@@ -25,25 +27,33 @@ public class CategoryServiceImpl implements CategoryService{
 
     @Override
     public Category getCategory(Long id) {
+        if(!categoryRepository.existsById(id))
+            throw new ResourceNotFoundException("Category", "CategoryId", id);
         return categoryRepository.getById(id);
     }
 
     @Override
-    public String addCategory(Category category) {
-        categoryRepository.save(category);
-         return "Added successfully !! ";
+    public Category addCategory(Category category) {
+        // adding duplicate need to handle
+        return categoryRepository.save(category);
     }
 
     @Override
-    public Category updateCategory(Category category, Long id) {
-        categoryRepository.save(category);
-        return null;
+    public Category updateCategory(Category updated, Long id) {
+        if(!categoryRepository.existsById(id))
+            throw new ResourceNotFoundException("Category", "CategoryId", id);
+        Category existing = categoryRepository.getById(id);
+        categoryAssambler.assamble(existing, updated);
+        return categoryRepository.save(existing);
     }
 
     @Override
     public String deleteCategory(Long id) {
+        if(categoryRepository.existsById(id))
+            throw new ResourceNotFoundException("Category", "CategoryId", id);
         categoryRepository.deleteById(id);
-        return "deleted !!";
+        return "deleted successfully !!";
+
 
     }
 }
